Rename package to privileged-service
