+++ src/go/privileged-service/pkg/svc/supervisor.go
@@ -23,7 +23,7 @@ import (
 	"golang.org/x/sys/windows/svc"
 	"golang.org/x/sys/windows/svc/debug"
 
-	"github.com/rancher-sandbox/rancher-desktop/src/go/supervisor/pkg/port"
+	"github.com/rancher-sandbox/rancher-desktop/src/go/privileged-service/pkg/port"
 )
 
 // supervisor implements service handler interface for
